LayoutTests/svg/custom/svg-parse-overflow-2.html

<html>
<body>
<script>
document.body.offsetTop;
if (window.testRunner)
    testRunner.dumpAsText();
</script>
<svg>
<rect width="478" height="-105000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000" stroke="#000000">
</svg>
</body>
