LayoutTests/svg/custom/svg-parse-overflow-2-expected.txt

CONSOLE MESSAGE: line 8: Error: Invalid value for <rect> attribute height="-105000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
